+++ src/vitest/index.d.ts
@@ -1,0 +1,3 @@
+export * from "./mock-console.js";
+export * from "./mock-local-storage.js";
+export * from "./setup-pinia.js";