+++ src/vue/index.d.ts
@@ -1,0 +1,2 @@
+export * from "./create-mount.js";
+export * from "./with-app-context.js";